requirements.txt
streamlit==1.28.1
openai==1.3.0
pinecone-client==3.0.0
rerankers[flashrank]==0.5.0
pandas==2.0.3